requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
jinja2==3.1.4
opencv-python==4.10.0.84
numpy==2.1.1
scikit-fuzzy==0.4.2
cryptography==43.0.1
pyjwt==2.9.0
SQLAlchemy==2.0.35
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
requests==2.32.3

